requirements.txt
fastapi
uvicorn
google-generativeai
python-docx
openpyxl
python-pptx
python-dotenv
